+++ latch_cli/utils/__init__.py
@@ -202,7 +202,8 @@ def hash_directory(dir_path: Path) -> str:
                     continue
 
                 exclude.append(l)
-    except FileNotFoundError: ...
+    except FileNotFoundError:
+        ...
 
     from docker.utils import exclude_paths